Start refactor tempo and time signatures handling
Before tempo and time signatures were statically passed to the
quantizer. But this turned out to be rather limiting. Its better
to directly pass tempo and time signature directly to the convert
method by attaching them directly to the mutwo event. This change
should happen slowly, therefore there are still the "default_.."
parameters.

Furthermore the quantizers don't use the tempo_envelope property,
but the builder does. So everything related to tempo has been moved
to the builder.