parameters: Refactor AbjadAttachment structure
The old structure of multi inheriting from a music_parameters.Indicator
and a AbjadAttachment class lead to the state that it was impossible
(or at least very ugly) to add new common attributes to all AbjadAttachments
which are not part of all indicators.

It may actually be necessary to pass to all 'AbjadAttachment' objects
common attributes, which don't matter for a music internal
representation. For this reason I refactored the internal structure