AbjadAttachment: Add MutwoToAbjad converters attrs
Various AbjadAttachment may need converters from mutwo to abjad world.
For instance the NaturalHarmonic attachment needs abjad pitches to
override the pitch segment of a note with the pitches the player needs
to see. So we actually need to assign an abjad pitch to a
PreciseNaturalHarmonic playing indicator. This violates the mutwo zen,
which suggestes software agnostic representations. To fix this we
can parse the mutwo_pitch_to_abjad_pitch converter, then we can also
assign mutwo pitches to the PreciseNaturalHarmonic indicator and only
during abjad conversion convert the mutwo pitches.