attachments: Add replace_leaf_by_leaf property
All abjad attachments process leaf-by-leaf currently. But sometimes it
may be more convenient to rather return a completely new abjad container
or structure (for instance for post-quantization). For this purpose the
new 'replace_leaf_by_leaf' property has been added to 'AbjadAttachment'.

Set to ``True`` if 'process_leaf_tuple' returns a sequence of
abjad leaves with an equal length to the provided leaf_tuple
and where each new leaf replaces the respective old leaf at
the same position (this is the default mode and was the only option
before this patch).

Set to ``True`` if 'process_leaf_tuple' returns a new abjad
object which is set at the position of the old first leaf and
if all other old leaves are supposed to be deleted.