converters/events/building: Explicitly let exception pass
If we quantize events with a duration line based quantizer, only the
first abjad leaf of a related leaf group is a note, and all following
events are skips. This is a necessary property to ensure that duration
line based notation works in lilypond. But this breaks the assumption of
the 'SequentialEventToAbjadVoice' converter, that all leaves of a leaf
group have the same leaf type. Because of this we sometimes had skips,
when notes were expected. Before this patch this was accepted, but the
problem with this code is, that it also hides problems/bugs, if the quantized
event accidentally misses note heads. By making the code more explicit
in which state it's ok to miss this, it's easier to track bugs.