converters/events: Refactor tuplet concatenation to dedicated function
Generally there is a law in mutwo 'use staticmethods, if utility
function is only needed in the given class, use utility module
otherwise' (the idea is not to flood the namespace with useless
functions). Due to this law, the tuplet concatenation was only part of
the quantizer class before this patch. But there are multiple good
reasons, why it may be better to move it to the utilities module:

1. Its functionality is actually a general extension of the abjad module
   and could be useful for anyone else using abjad.

2. It's a very complex function and it becomes easier to read and to
   test if it exists in a dedicated module.