converters/SequentialEventToAbjadVoice: Set default quantizer to LeafMakerSequentialEventToQuantizedAbjadContainer
Before this patch the default quantizer of 'SequentialEventToAbjadVoice'
was 'NauertSequentialEventToQuantizedAbjadContainer'. This quantizer was
the first quantizer, that was added to mutwo. But nowadays I use much
more often 'LeafMakerSequentialEventToQuantizedAbjadContainer' as it has
some strong advantages:

- it's much faster
- it's deterministic
- it's much simpler & therefore easier to debug

Because I think those advantages are more important for an average user
than being able to represent very complex rhythm (although it's not even
always sure that the Nauert based quantizer represents them correctly),
this patch changes the default quantizer.

On the long term, it also seems more likely to maintain the leaf maker
based quantizer than the Nauert based quantizer due to mentioned
advantages.