converters/events: Mv tie + is_simple_event_rest to quantizer
In this way 'SequentialEventToAbjadVoice' becomes more flexible and it's
no longer per-se necessary that mutwo data that's converted to abjad
data can't add indicators to two adjacent rests. This is now only a
limitation of the currently implemented quantizers and not of the
general 'SequentialEventToAbjadVoice' converter.

In this way the quantizer can also be responsible itself for adding
'is_rest' attributes to the user provided simple events, and they no
longer need to rely on a outside code ('SequentialEventToAbjadVoice')
that this happens.