utilities/tests: Make tests more reliable and faster
Before this patch unit tests compared lilypond-rendered images and
checked if they are still the same. This approach had multiple issues:

(1) It was very slow as we always had to wait Lilypond to render and the
    png comparison is also expensive.

(2) It was unreliable as minimal changes between two Lilypond version
    that minimally changed the output image resulted in failed tests,
    although the semantics hadn't changed at all.

Particular (2) was a big problem, as tests could only run successfully
if we had exactly the same lilypond version on the machine as the one
where the test files were created. By comparing lilypond text files
instead of rendered png files we can fix both problems and make tests
faster and more reliable.